requirements.txt
fastapi
uvicorn[standard]
streamlit>=1.24
datasets>=2.12
sentence-transformers>=2.2
faiss-cpu>=1.7.4
numpy
pandas
tqdm
scikit-learn
transformers>=4.30
openai
python-dotenv
pytest
requests
rank-bm25
sentencepiece
fuzzywuzzy[speedup]
